extract food removal functions to diary-food-remove
Extracts all food removal functions to diary-food-remove.js.
Necessary changes made to diary.js to allow for appropriate function
of app features. All page behavior is normal.
