+++ lib/diary/diary.js
@@ -3,6 +3,7 @@ const LOCAL = 'http://localhost:3000/api/v1/'
 const diaryFoodTable = require('./diary-food-table')
 const calories = require('./calories')
 const foodAdder = require('./diary-food-add')
+const foodRemover = require('./diary-food-remove')
 $ = require('jquery')
 
 let counter = 0
@@ -140,30 +141,6 @@ function searchFoods(){
   })
 }
 
-function removeFoodFromMealTable(button) {
-  let mealName = $(button).offsetParent().find('span').text()
-  let calorieUpdate = -parseInt($(button).parent().prev().text())
-  $(button).parent().parent().remove()
-
-  calories.updateMealTotalCalories(mealName, {calories: calorieUpdate})
-  calories.updateMealRemainingCalories(mealName, {calories: calorieUpdate})
-  calories.updateRemainingCalories()
-  calories.updateConsumedCalories()
-}
-
-function removeFoodFromMeal(){
-  const button = event.target
-  const id = $(button).data("meal-id")
-
-  return $.ajax({
-    url: `${API}meals/${id}`,
-    method: 'DELETE'
-  })
-  .then(removeFoodFromMealTable(button))
-  .fail((error) => {
-    console.error(error)
-  })
-}
 $(document).ready(function(){
   const diary = new Diary
 
@@ -188,7 +165,7 @@ $(document).ready(function(){
     foodAdder.addMeal()
   })
   $('body').on('click','.remove', (event) => {
-    removeFoodFromMeal()
+    foodRemover.removeFoodFromMeal()
   })
   $('body').on('click', '#prev-page-btn', (event) => {
     counter = counter - 1
